Working visualization within QT GUI
